requirements.txt
amqp==5.0.6
APScheduler==3.7.0
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-apscheduler==0.5.2
django-filter==2.4.0
idna==2.10
kombu==5.1.0
oauthlib==3.1.1
prompt-toolkit==3.0.19
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
